Allow owner to nickname animals
Add tests

+++ test/TestAdoption.sol
@@ -6,54 +6,78 @@ import "truffle/DeployedAddresses.sol";
 import "../contracts/Adoption.sol";
 
 contract TestAdoption {
-  // The address of the adoption contract to be tested
-  Adoption adoption = Adoption(DeployedAddresses.Adoption());
+    // The address of the adoption contract to be tested
+    Adoption adoption = Adoption(DeployedAddresses.Adoption());
 
-  // The id of the animal that will be used for testing
-  uint expectedAnimalId = 8;
+    // The id of the animal that will be used for testing
+    uint256 expectedAnimalId = 8;
 
-  //The expected owner of adopted animal is this contract
-  address expectedAnimalAdopter = address(this);
+    //The expected owner of adopted animal is this contract
+    address expectedAnimalAdopter = address(this);
 
-  // Testing the adoptAnimal() function
-  function testUserCanAdoptAnimal() public {
-    uint adoptedAnimalId = adoption.adoptAnimal(expectedAnimalId);
+    // Testing the adoptAnimal() function
+    function testUserCanAdoptAnimal() public {
+        uint256 adoptedAnimalId = adoption.adoptAnimal(expectedAnimalId);
 
-    Assert.equal(adoptedAnimalId, expectedAnimalId, "adoptAnimal should return the id of the expected animal");
-  }
+        Assert.equal(
+            adoptedAnimalId,
+            expectedAnimalId,
+            "adoptAnimal should return the id of the expected animal"
+        );
+    }
 
-  // Testing retrieval of a single animal's owner
-  function testGetAdopterAddressByAnimalId() public {
-    address animalAdopter = adoption.animalAdopters(expectedAnimalId);
+    // Testing retrieval of a single animal's owner
+    function testGetAdopterAddressByAnimalId() public {
+        address animalAdopter = adoption.animalAdopters(expectedAnimalId);
 
-    Assert.equal(animalAdopter, expectedAnimalAdopter, "Owner of the expected animal should be this contract");
-  }
+        Assert.equal(
+            animalAdopter,
+            expectedAnimalAdopter,
+            "Owner of the expected animal should be this contract"
+        );
+    }
 
-  // Testing retrieval of all animal owners
-  function testGetAdopterAddressByAnimalIdInArray() public {
-    // Store adopters in memory rather than contract's storage
-    address[16] memory animalAdopters = adoption.getAnimalAdopters();
+    // Testing retrieval of all animal owners
+    function testGetAdopterAddressByAnimalIdInArray() public {
+        // Store adopters in memory rather than contract's storage
+        address[16] memory animalAdopters = adoption.getAnimalAdopters();
 
-    Assert.equal(animalAdopters[expectedAnimalId], expectedAnimalAdopter, "Owner of the expected animal should be this contract");
-  }
+        Assert.equal(
+            animalAdopters[expectedAnimalId],
+            expectedAnimalAdopter,
+            "Owner of the expected animal should be this contract"
+        );
+    }
 
-  // Testing the returnAnimal() function
-  function testUserCanReturnAnimal() public {
-    uint returnedAnimalId = adoption.returnAnimal(expectedAnimalId);
+    // Testing the returnAnimal() function
+    function testUserCanReturnAnimal() public {
+        uint256 returnedAnimalId = adoption.returnAnimal(expectedAnimalId);
 
-    Assert.equal(returnedAnimalId, expectedAnimalId, "returnAnimal should return the id of the expected animal");
-  }
+        Assert.equal(
+            returnedAnimalId,
+            expectedAnimalId,
+            "returnAnimal should return the id of the expected animal"
+        );
+    }
 
-  // Testing retrieval of a single animal's owner
-  function testGetReturnerAddressByAnimalId() public {
-    address animalAdopter = adoption.animalAdopters(expectedAnimalId);
+    // Testing retrieval of a single animal's owner
+    function testGetReturnerAddressByAnimalId() public {
+        address animalAdopter = adoption.animalAdopters(expectedAnimalId);
 
-    Assert.equal(animalAdopter,  address(0x0), "Owner of the expected animal is reset to the default address");
-  }
+        Assert.equal(
+            animalAdopter,
+            address(0x0),
+            "Owner of the expected animal is reset to the default address"
+        );
+    }
 
-  function testGetReturnerAddressByAnimalIdInArray() public {
-    // Store adopters in memory rather than contract's storage
-    address[16] memory animalAdopters = adoption.getAnimalAdopters();
-    Assert.equal(animalAdopters[expectedAnimalId], address(0x0), "Owner of the expected animal is reset to the default address");
-  }
+    function testGetReturnerAddressByAnimalIdInArray() public {
+        // Store adopters in memory rather than contract's storage
+        address[16] memory animalAdopters = adoption.getAnimalAdopters();
+        Assert.equal(
+            animalAdopters[expectedAnimalId],
+            address(0x0),
+            "Owner of the expected animal is reset to the default address"
+        );
+    }
 }
